requirements.txt
pandas
numpy
psycopg2-binary
apache-airflow[postgres]
jupyter
beautifulsoup4
requests
sqlalchemy
ipdb
matplotlib